treat free list as a stack rather than a queue?
Rather than treat the free list as a queue, where a buffer is
unshifted from the head of the list during Get and shifted to the tail
of the list during a Put, treat the free list as a stack, because it's
more quick to pop off the item from the tail of the list than to shift
the list by a single element.

Or so the thinking goes.  I have added that benchmark to show that
working with a slice as a stack is twice the performance of a queue,
when that logic is placed into the gobp2 structure, the resulting
benchmarks for that data structure are not as strong as for the
original gobp structure.